feat: handle history ui changes for request fulfillment using bank accounts
