+++ src/components/TransactionDetails/transactionTransformer.ts
@@ -57,6 +57,8 @@ export interface TransactionDetails {
         isLinkTransaction?: boolean
         transactionCardType?: TransactionCardType
         rewardData?: RewardData
+        fulfillmentType?: 'bridge' | 'wallet'
+        bridgeTransferId?: string
         depositInstructions?: {
             amount: string
             currency: string
@@ -173,14 +175,22 @@ export function mapTransactionDataForDrawer(entry: HistoryEntry): MappedTransact
             }
             break
         case EHistoryEntryType.REQUEST:
-            if (entry.userRole === EHistoryUserRole.RECIPIENT) {
+            if (entry.extraData?.fulfillmentType === 'bridge' && entry.userRole === EHistoryUserRole.SENDER) {
+                transactionCardType = 'bank_request_fulfillment'
+                direction = 'bank_request_fulfillment'
+                nameForDetails = entry.recipientAccount?.username ?? entry.recipientAccount?.identifier
+                isPeerActuallyUser = !!entry.recipientAccount?.isUser || !!entry.senderAccount?.isUser
+            } else if (entry.userRole === EHistoryUserRole.RECIPIENT) {
                 direction = 'request_sent'
                 transactionCardType = 'request'
                 nameForDetails =
                     entry.senderAccount?.username || entry.senderAccount?.identifier || 'Requested via Link'
                 isPeerActuallyUser = !!entry.senderAccount?.isUser
             } else {
-                if (entry.status?.toUpperCase() === 'NEW' || entry.status?.toUpperCase() === 'PENDING') {
+                if (
+                    entry.status?.toUpperCase() === 'NEW' ||
+                    (entry.status?.toUpperCase() === 'PENDING' && !entry.extraData?.fulfillmentType)
+                ) {
                     direction = 'request_received'
                     transactionCardType = 'request'
                     nameForDetails =
@@ -252,7 +262,8 @@ export function mapTransactionDataForDrawer(entry: HistoryEntry): MappedTransact
     if (
         entry.type === EHistoryEntryType.BRIDGE_OFFRAMP ||
         entry.type === EHistoryEntryType.BRIDGE_ONRAMP ||
-        entry.type === EHistoryEntryType.BANK_SEND_LINK_CLAIM
+        entry.type === EHistoryEntryType.BANK_SEND_LINK_CLAIM ||
+        entry.extraData?.fulfillmentType === 'bridge'
     ) {
         switch (entry.status?.toUpperCase()) {
             case 'AWAITING_FUNDS':
@@ -363,8 +374,12 @@ export function mapTransactionDataForDrawer(entry: HistoryEntry): MappedTransact
             isLinkTransaction: isLinkTx,
             transactionCardType,
             rewardData,
+            fulfillmentType: entry.extraData?.fulfillmentType,
+            bridgeTransferId: entry.extraData?.bridgeTransferId,
             depositInstructions:
-                entry.type === EHistoryEntryType.BRIDGE_ONRAMP ? entry.extraData?.depositInstructions : undefined,
+                entry.type === EHistoryEntryType.BRIDGE_ONRAMP || entry.extraData?.fulfillmentType === 'bridge'
+                    ? entry.extraData?.depositInstructions
+                    : undefined,
             receipt: entry.extraData?.receipt,
         },
         sourceView: 'history',
