+++ src/services/services.types.ts
@@ -205,27 +205,6 @@ export interface TRequestChargeResponse {
     }
 }
 
-// create payment response
-export interface PaymentResponse {
-    uuid: string
-    paidTokenAddress: string
-    payerChainId: string
-    payerTransactionHash: string
-    requestCharge: {
-        uuid: string
-        chainId: string
-        createdAt: string
-        tokenAddress: string
-        tokenAmount: string
-        tokenDecimals: number
-        requestLink: {
-            recipientAddress: string
-            reference?: string
-            attachmentUrl?: string
-        }
-    }
-}
-
 enum EHistoryEntryType {
     CLAIM = 'CLAIM',
     REQUEST = 'REQUEST',
